compiling, but in correct, solver complete

+++ src/main/java/net/antineutrino/SudokuAlgos/ArrayUtils.java
@@ -1,7 +1,7 @@
 package net.antineutrino.SudokuAlgos;
 
 public class ArrayUtils{
-	public static int findIntIndex(int[] array, int val){
+	public static int findIntIndex(byte[] array, byte val){
 		for (int i=0; i < array.length; i++){
 			if (array[i] == val) {
 				return i;
